+++ scripts/release.sh
@@ -1,48 +1,43 @@
 #!/usr/bin/env bash
 set -e
 
-current_branch="$(git rev-parse --abbrev-ref HEAD)"
-release_branch="${1:-production}"
+current_branch="$(git rev-parse --abbrev-ref HEAD)";
+release_branch="${1:-production}";
 
 if [[ "$release_branch" == "production" && "$current_branch" != "staging" ]]; then
-  echo "Releasing production can only be done while on staging branch"
+  echo "Releasing production can only be done while on staging branch";
   exit 1;
 elif [[ "$release_branch" == "$current_branch" ]]; then
-  echo "Releasing $release_branch can't be done while on $current_branch branch"
-  echo "Make sure to checkout to your feature branch"
+  echo "Releasing $release_branch can't be done while on $current_branch branch";
+  echo "Make sure to checkout to your feature branch";
   exit 1;
 fi
 
+echo "Updating $release_branch";
+git checkout "$release_branch";
+git pull origin "$release_branch";
+
 if [[ "$release_branch" == "alpha" || "$release_branch" == "beta" ]]; then
-  echo "Deleting local and remote $release_branch branch"
-  git branch -D "$release_branch" || true
-  git push origin --delete "$release_branch" || true
+  echo "Merging changes";
+  git merge -X theirs "$current_branch";
 
-  echo "Checking out and pushing a clean $release_branch branch"
-  git checkout -b "$release_branch";
-  git push origin "$release_branch"
-fi
+  echo "Running release-local on affected projects";
+  pnpm nx affected --target=release-local --base=staging --parallel=false;
+else
+  git checkout "$current_branch";
 
-echo "Running build in parallel to improve performance"
-pnpm nx affected --target=build --parallel=6;
+  echo "Running release-local on affected projects";
+  pnpm nx affected --target=release-local --base="$release_branch" --parallel=false;
 
-echo "Running release-local on affected projects"
-pnpm nx affected --target=release-local --parallel=false;
-
-if [[ "$release_branch" == "production" && "$current_branch" == "staging" ]]; then
-  echo "Checking out to $release_branch"
+  echo "Checking out to $release_branch";
   git checkout "$release_branch";
 
-  echo "Updating branch"
-  git pull;
-
-  echo "Merging changes"
+  echo "Merging changes";
   git merge --ff-only "$current_branch";
+
+  echo "Pushing changes to $release_branch";
+  git push origin "$release_branch";
 fi
 
-echo "Pushing changes to $release_branch"
-git push -u origin "$release_branch";
-
-echo "Merge changes back to $current_branch"
-git checkout "$current_branch"
-git merge --ff-only "$release_branch"
+echo "Go back to $current_branch";
+git checkout "$current_branch";